+++ fwdpp/fwd_functional.hpp
@@ -4,6 +4,8 @@
 #include <cmath>
 #include <limits>
 #include <functional>
+#include <gsl/gsl_rng.h>
+#include <gsl/gsl_randist.h>
 
 /*! \file fwd_functional.hpp
   Defines several function objects used both internally and by library users
@@ -49,6 +51,44 @@ namespace KTfwd
         operator()(const mtype &m) const
         {
             return m.neutral;
+        }
+    };
+
+    struct poisson_interlocus_rec
+    /// \brief Model inter-locus/region recombination as a Poisson process
+    /// \ingroup mlocus
+    {
+        /// The number of cross-overs between loci i and i+1
+        /// are Poisson-distributed with this mean
+        const double mean;
+
+        /// Construct with a mean value.
+        explicit poisson_interlocus_rec(const double mean_) : mean(mean_) {}
+
+        inline unsigned
+        operator()(const gsl_rng *r) const noexcept
+        /// \return Number of crossovers between loci i and i+1
+        {
+            return gsl_ran_poisson(r, mean);
+        }
+    };
+
+    struct binomial_interlocus_rec
+    /// \brief Model inter-locus/region recombination as a Binomial process.
+    /// \ingroup mlocus
+    {
+		/// Genetic distance between loci i and i+1 is in centiMorgans (cM).
+        const double cM;
+
+		/// Construct with number of cM
+        explicit binomial_interlocus_rec(const double cM_) : cM(cM_) {}
+
+        inline unsigned
+        operator()(const gsl_rng *r) const noexcept
+		/// Return whether or not loci i and i+1 had an odd number of crossovers
+		/// between them.
+        {
+            return gsl_ran_binomial(r, cM, 1);
         }
     };
 }