+++ doc/md/RELEASE_NOTES.md
@@ -6,6 +6,7 @@ milestones (target version numbers for these features to go live) because that i
 
 ## 0.5.6
 
+* API issue #48 is addressed.  This is an API change breaking backwards compatilibity for multi-locus/region simulations. The new API allows more flexibility in modeling interlocus/region crossovers.  The API for KTfwd::sample_diploid is changed for multi-locus/region sims.  The new types KTfwd::poisson_interlocus_rec and KTfwd::binomial_interlocus_rec are provided to facilitate modeling interlocus recombination.  Two functions, KTfwd::make_poisson_interlocus_rec and KTfwd::make_binomial_interlocus_rec, return vectors of callbacks bound to the two new structs.
 * API issue #49 is addressed.  The changes maintain compile-time API compatibility with previous library versions.
 * Extend concept of regions to multi-locus sims via extensions::bind_vec__drm and extensions::bind_vec::dmm. [commit](https://github.com/molpopgen/fwdpp/commit/da1b47b661679c80530b3ed477107f9fadf33e25)
 * extensions::discrete_mut_model::make_mut now takes a pointer to the generation, allowing a single point of binding rather than binding each generation. [commit](https://github.com/molpopgen/fwdpp/commit/da1b47b661679c80530b3ed477107f9fadf33e25)